Cl button (#1095)
* cl button

* shared button

* shared button
