--- src/applications/content-entries-app/entry/components/content-lab-button/content-lab-button.component.html
@@ -1,22 +1,0 @@
-<button pButton class="kButtonBranded"
-        [kTooltip]="reason ? (('contentLab.errors.' + reason) | translate) : ''"
-        [label]="loading ? ('contentLab.initializing' | translate) : ('contentLab.buttonLabel' | translate)"
-        [disabled]="disabled"
-        [ngClass]="loading ? 'loading' : disabled ? 'disabled' : 'loaded'"
-        [icon]="'clIcon'"
-        (click)="openContentLab()"></button>
-
-<kPopupWidget #clipAndTrim [fullScreen]="true">
-    <ng-template>
-        <kEntryEditor [entry]="_contentLabSelectedQuiz"
-                      [entryHasSource]="true"
-                      [tab]="_isQuizEntry? 'quiz' : 'editor'"
-                      [parentPopupWidget]="clipAndTrim"></kEntryEditor>
-    </ng-template>
-</kPopupWidget>
-
-<kPopupWidget #bulkActionsPopup [popupWidth]="570" [popupHeight]="245" [modal]="true" [preventPageScroll]="true">
-    <ng-template>
-        <kDownloadEntry [entry]="_entryStore.entry" [parentPopupWidget]="bulkActionsPopup"></kDownloadEntry>
-    </ng-template>
-</kPopupWidget>
