+++ src/shared/content-shared/entries/entries-list/entries-list.component.html
@@ -98,6 +98,7 @@
                      [sortOrder]="_query.sortDirection"
                      [columns]="columns"
                      [rowActions]="rowActions"
+                     [isAIButtonVisible]="isAIButtonVisible"
                      (actionSelected)="onActionsSelected.emit($event)"
                      (openActionsMenu)="_openActionsMenu($event)"
                      (sortChanged)="onSortChanged($event)"
