+++ src/shared/kmc-shared/kmc-views/component-views/advertisements-app-view.service.ts
@@ -56,7 +56,7 @@ export class AdvertisementsAppViewService extends KmcComponentViewBaseService<Ad
     private _isAvailableByData(args: AdvertisementsAppViewArgs): boolean {
         const { entry, hasSource} = args;
         const entryReady = entry.status === KalturaEntryStatus.ready;
-        const isEntryReplacing = entry.replacementStatus !== KalturaEntryReplacementStatus.none;
+        const isEntryReplacing = entry.replacementStatus && entry.replacementStatus !== KalturaEntryReplacementStatus.none;
         const isLiveEntry = entry.mediaType === KalturaMediaType.liveStreamFlash ||
             entry.mediaType === KalturaMediaType.liveStreamWindowsMedia ||
             entry.mediaType === KalturaMediaType.liveStreamRealMedia ||
