+++ src/shared/kmc-shared/kmc-views/component-views/clip-and-trim-app-view.service.ts
@@ -58,7 +58,7 @@ export class ClipAndTrimAppViewService extends KmcComponentViewBaseService<ClipA
     private _isAvailableByData(args: ClipAndTrimAppViewArgs): boolean {
         const { entry, hasSource} = args;
         const entryReady = entry.status === KalturaEntryStatus.ready;
-        const isEntryReplacing = entry.replacementStatus !== KalturaEntryReplacementStatus.none;
+        const isEntryReplacing = entry.replacementStatus && entry.replacementStatus !== KalturaEntryReplacementStatus.none;
         const isExternalMedia = entry instanceof KalturaExternalMediaEntry;
         const isEntryRelevant = [KalturaMediaType.video, KalturaMediaType.audio].indexOf(entry.mediaType) !== -1 && !isExternalMedia;
         const isLiveEntry = [
